fixed Schrodinger's licence check

+++ DockTbx/glide.py
@@ -4,9 +4,13 @@ import glob
 import shutil
 import subprocess
 
-import method
+import numpy as np
 
-from DockTbx.licence import check as chkl
+from DockTbx import method
+
+from DockTbx.tools import reader
+from DockTbx.tools import mol2
+from DockTbx.license import check as chkl
 
 required_programs = ['prepwizard', 'glide', 'ligprep', 'glide_sort', 'pdbconvert']
 
@@ -43,15 +47,11 @@ class Glide(method.DockingMethod):
         locals().update(self.options)
 
         # prepare protein cmd (the protein structure is already assumed to be minimized/protonated with prepwizard)
-        prepwizard_cmd = chkl.eval("prepwizard -WAIT -fix %(file_r)s target.mae"%locals(), 'schrodinger')    
-        #prepwizard_cmd = chkl.eval("prepwizard -WAIT -noprotassign -nohtreat -noimpref %(file_r)s target.mae"%locals(), 'schrodinger')    
+        prepwizard_cmd = chkl.eval("prepwizard -fix %(file_r)s target.mae"%locals(), 'schrodinger')    
 
-        # prepare grid cmd
-        glide_grid_cmd = chkl.eval("glide -WAIT grid.in", 'schrodinger') # grid prepare
-        structconvert_cmd = chkl.eval("structconvert -imol2 %(file_l)s -omae lig.mae"%locals(), 'schrodinger')
-    
-        # docking cmd
-        glide_dock_cmd = chkl.eval("glide -WAIT dock.in", 'schrodinger') # docking command
+        # prepare grid and docking cmd
+        glide_grid_cmd = chkl.eval("glide grid.in", 'schrodinger')
+        glide_dock_cmd = chkl.eval("glide dock.in", 'schrodinger')
 
         tmpdirline = self.tmpdirline
     
@@ -77,7 +77,7 @@ RECEP_FILE target.mae" > grid.in
 %(glide_grid_cmd)s
 
 # (C) convert ligand to maestro format
-%(structconvert_cmd)s
+structconvert -imol2 %(file_l)s -omae lig.mae
 
 # (D) perform docking
 echo "WRITEREPT YES
@@ -94,29 +94,28 @@ PRECISION %(precision)s" > dock.in
     def extract_docking_results(self, file_s, input_file_r, input_file_l):
         """Extract Glide docking results""" 
 
-        # cmd to extract results
-        glide_sort_cmd = chkl.eval("glide_sort -r sort.rept dock_pv.maegz -o dock_sorted.mae", 'schrodinger', redirect='/dev/null')
-        subprocess.check_output(glide_sort_cmd, shell=True, executable='/bin/bash')
+        if os.path.exists('dock_pv.maegz'):
+            # (1) cmd to extract results
+            subprocess.check_output('glide_sort -r sort.rept dock_pv.maegz -o dock_sorted.mae', shell=True, executable='/bin/bash')
 
-        # convert to .mol2
-        convert_cmd = chkl.eval("mol2convert -n 2: -imae dock_sorted.mae -omol2 dock_sorted.mol2", 'schrodinger', redirect='/dev/null')
-        subprocess.check_output(convert_cmd, shell=True, executable='/bin/bash')
+            # (2) convert to .mol2
+            subprocess.check_output('mol2convert -n 2: -imae dock_sorted.mae -omol2 dock_sorted.mol2', shell=True, executable='/bin/bash')
 
-        # create multiple files with babel
-        subprocess.check_output('babel -imol2 dock_sorted.mol2 -omol2 lig-.mol2 -m &>/dev/null', shell=True, executable='/bin/bash')
-
-        # extract scores
-        with open('dock.rept', 'r') as ffin:
-            with open(file_s, 'w') as ffout:
-                line = ffin.next()
-                while not line.startswith('===='):
-                    line = ffin.next()
-                while True:
-                    line = ffin.next()
-                    if line.strip():
-                        print >> ffout, line.split()[3]
-                    else:
-                        break
+            if os.path.exists('dock_sorted.mol2'):
+                ligname = reader.open(input_file_l).ligname
+                mol2.update_mol2file('dock_sorted.mol2', 'lig-.mol2', ligname=ligname, multi=True)
+                # extract scores
+                with open('dock.rept', 'r') as ffin:
+                    with open(file_s, 'w') as ffout:
+                        line = ffin.next()
+                        while not line.startswith('===='):
+                            line = ffin.next()
+                        while True:
+                            line = ffin.next()
+                            if line.strip():
+                                print >> ffout, line[43:51].strip()
+                            else:
+                                break
 
     def get_tmpdir_line(self):
         if self.options['tmpdir']:
@@ -124,33 +123,30 @@ PRECISION %(precision)s" > dock.in
         else:
             line = ""
 
-    def write_rescoring_script(self, filename, file_r, file_l):
+    def write_rescoring_script(self, filename, file_r, files_l):
         """Rescore using Glide SP scoring function"""
         locals().update(self.options)
 
+        files_l_joined = ' '.join(files_l)
+
         # prepare protein cmd (the protein structure is already assumed to be minimized/protonated with prepwizard)
-        prepwizard_cmd = chkl.eval("prepwizard -WAIT -noprotassign -nohtreat -noimpref %(file_r)s target.mae"%locals(), 'schrodinger')
+        prepwizard_cmd = chkl.eval("prepwizard -fix %(file_r)s target.mae"%locals(), 'schrodinger')
 
-        # prepare grid cmd
-        glide_grid_cmd = chkl.eval("glide -WAIT grid.in", 'schrodinger') # grid prepare
-        structconvert_cmd = chkl.eval("structconvert -imol2 %(file_l)s -omae lig.mae"%locals(), 'schrodinger')
-
-        # scoring cmd
-        glide_dock_cmd = chkl.eval("glide -WAIT dock.in", 'schrodinger') # docking command
+        # prepare grid and scoring cmd
+        glide_grid_cmd = chkl.eval("glide grid.in", 'schrodinger') # grid prepare
+        glide_dock_cmd = chkl.eval("glide dock.in", 'schrodinger') # docking command
         tmpdirline = self.tmpdirline
 
         with open(filename, 'w') as file:
             script ="""#!/bin/bash
 %(tmpdirline)s
+cat %(files_l_joined)s > lig.mol2
 
-if [ ! -f target.mae ]; then
-  # (A) Prepare receptor
-  %(prepwizard_cmd)s
-fi
+# (A) Prepare receptor
+%(prepwizard_cmd)s
 
-if [ ! -f grid.zip ]; then
-  # (B) Prepare grid
-  echo "USECOMPMAE YES
+# (B) Prepare grid
+echo "USECOMPMAE YES
 INNERBOX %(innerbox)s
 ACTXRANGE %(actxrange)s
 ACTYRANGE %(actyrange)s
@@ -160,11 +156,11 @@ OUTERBOX %(outerbox)s
 ENTRYTITLE target
 GRIDFILE grid.zip
 RECEP_FILE target.mae" > grid.in
-  %(glide_grid_cmd)s
-fi
+%(glide_grid_cmd)s
+
 
 # (C) convert ligand to maestro format
-%(structconvert_cmd)s
+structconvert -imol2 lig.mol2 -omae lig.mae
 
 # (D) perform rescoring
 echo "WRITEREPT YES
@@ -176,16 +172,28 @@ PRECISION SP" > dock.in
 
 %(glide_dock_cmd)s"""% locals()
             file.write(script)
-    
-    def extract_rescoring_results(self, filename):
-    
-        with open(filename, 'a') as ff:
-            with open('dock.scor', 'r') as outf:
-                for line in outf:
-                    if line.startswith('   1'):
-                        print >> ff, line.split()[3]
-    
+
+    def extract_rescoring_results(self, filename): 
+        idxs = []
+        scores = []
+
+        with open('dock.scor', 'r') as ffin:
+            line = ffin.next()
+            while not line.startswith('===='):
+                line = ffin.next()
+            while True:
+                line = ffin.next()
+                if line.strip():
+                    idxs.append(int(line[36:42].strip()))
+                    scores.append(line[43:51].strip())
+                else:
+                    break
+
+        scores = np.array(scores)
+        scores = scores[np.argsort(idxs)]
+        with open(filename, 'w') as ffout:
+            for sc in scores:
+                print >> ffout, sc
+
     def cleanup(self):
         pass
-        #for ff in glob.glob('dock_sorted*.pdb'):
-        #    os.remove(ff)
